Moved Components into subfolders
